requirements.txt
PyJWT
passlib
asgiref==3.3.4
click==8.0.1
fastapi==0.65.1
h11==0.12.0
mongoengine==0.23.1
pydantic==1.8.2
pymongo==3.11.4
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
